MGMT-15578: Collect event logs and cluster logs from failed MCE clusters (#54)
This change introduces the capability to remove service and event logs for failed MCE agents.

Related Issue: https://issues.redhat.com/browse/MGMT-15578

Description of Changes:
We are adding an ability to the MCE gather to obtain logs and events from agentclusterinstall CR's that are marked as "failed" where these logs are available.

What resource is being added: | N/A

Is this a Hub or Managed cluster change?:
Hub Cluster | Managed Cluster | N/A

Notes:

Signed-off-by: Paul Maidment <pmaidmen@redhat.com>